Small changes in models

+++ users/models.py
@@ -48,7 +48,7 @@ class CustomUser(AbstractUser):
 class SalaryParams(models.Model):
     user = models.ForeignKey(CustomUser, on_delete=models.SET_NULL, null=True)
     fixed = models.PositiveIntegerField()
-    kpi_by_sales = models.PositiveIntegerField()
+    kpi_by_sales = models.FloatField()
     comment = models.TextField(null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(null=True, blank=True, auto_now=True)
@@ -57,7 +57,7 @@ class SalaryParams(models.Model):
 class SalaryPayment(models.Model):
     user = models.ForeignKey(CustomUser, on_delete=models.SET_NULL, null=True, related_name="user_salary_payments")
     payer = models.ForeignKey(CustomUser, on_delete=models.SET_NULL, null=True, related_name="payer_payments")
-    amount = models.PositiveIntegerField()
+    amount = models.FloatField()
     paid_at = models.DateTimeField()
     comment = models.TextField(null=True, blank=True)
